TS: Report failures (exceptions) using a file
We have not setup an http server yet. Thus we cannot communicate to the
client that a failure occured using the http server. This approach saves
the error to a world readable file. It's up to the client to read the
failure in that particular file. The file is saved in the 'files'
directory of the installed test-server.
e.g./data/data/com.example.test/files/calabash_failure.out.