Query: Avoid loading previously unloaded classes
This fixes an issue where querying for a qualified class name
would load the class in the classloader and call its static
initializers. E.g. query("my.qualified.class").

This would often happen when querying for a
specialized/custom viewtype as the first step of the test, as
it would interfere with the default lifecycle of the
application.