Problem: Webview queries report wrong coordinates
Solution: Add a comparison node at the absolute position 0,0. The
scrolling that is taken into account within the webview is offset by
this. Thus we get the coordinate of the html views in relative values to
the current viewport.

In the test-server, we then offset these relative values with the
absolute scroll of the webview, which includes the HTML scrolling.