Problem: QueryCaching: We should not cache ClassName queries
Problem
We cache the optimization we have made of a query. E.g. a String is
cached to the optimized UIQueryAST.
    A query for a class name will NOT load the class if it is not
already loaded. If the Class is loaded later on, then it will be used.
    If we continously query for a class before it is loaded, then, after
it is loaded, the query will be cached and not recognize & update to
search for the class.

Solution
We simply stop caching a query if it has fully qualified classes in it.

Issue
https://github.com/calabash/calabash-android-server/issues/31