Allow actions on parent when using ui automator
Background:

While introducing Jetpack compose in our android project, we noticed that calabash is not able to identify compose ui components while ui automator is. Even if ui automator is able to "see" elements in the ui, it is very difficult to interact with them because there are no identifiers. We have to use text and content description to identify components and interact with them.

Some compose components are identified by ui automator as multiple nodes and this prevents us from properly interacting with them. For example, one checkbox is represented by a parent node that is checkable and two child nodes: one containing the text and one containing the checkbox icon (both of them not checkable). As we can only use text to identify a component, in this case we receive the child node containing the text, that is not checkable. In order to access the "checked" status we need to get the parent node of the identified node and to be able to execute ui automator actions on it.

Solution:
Allow calabash to execute ui automator actions on the parent of a located node.