uiautomator_set_text command added
There was no way to set text using ui automator, in this PR we are adding the command uiautomator_set_text that will allow setting text into an input field and will verify that it worked and fail otherwise.

To be able to set the text to an input field in compose, calling `setText` doesn't work. For it to work we need to call `setText` in the parent of the text field element (the same issue we explained in https://github.com/calabash/calabash-android-server/pull/110. For this reason the command uiautomator_set_text allows a parameter to execute the method in the parent of the target node.