UiScrollableCustom modified to allow to force scroll into an element
New method added to make sure we keep compatibility with existing code. The new method allows to scroll into a view forcing the scroll even if the method scrollForward returns false (that means that the end of the scrollable view has been found). We discovered that the method scrollForward is unreliable and in some occasions returns a false while there is still space in the view to scroll to.