Scroll to element action improved
Scroll to element allows to scroll to an element providing a strategy, a locator and the index of the element (to be able to scroll to the second element meeting certain conditions for example). It allows to provide the index of the scrollView we want to use in case our screen has multiple scrollviews. Finally, it also allows to customise the maximum number of scrolls that we want to perform so we don't have the test trying to scroll for too much time.

Co-authored-by: Manjunatha Thimmaiah <manjunatha.tec@gmail.com>