Scrolling actions fixed
Two actions have been added: one for vertical and one for horizontal scrolling. Both of them receive the following parameters: strategy and locator to find the target view, optionally, the strategy and locator to find the container view and finally the max number of scrolls that we want to do before stop trying to find the element and returning an error.